+++ cdap-distributions/src/hdinsight/pkg/install.sh
@@ -102,7 +102,7 @@ sed \
 chef-solo -o 'recipe[ulimit::default],recipe[cdap::fullstack],recipe[cdap::init]' -j ${__tmpdir}/generated-conf.json || die 'Failed during Chef run'
 
 # Temporary Hack to workaround CDAP-4089
-rm -f /opt/cdap/kafka/lib/log4j.log4j-1.2.14.jar
+rm -f /opt/cdap/kafka-service/lib/log4j.log4j-1.2.14.jar
 
 # Start CDAP Services
 for i in /etc/init.d/cdap-*; do