+++ cdap-master/src/main/java/co/cask/cdap/MiscTest.java
@@ -1,0 +1,8 @@
+package co.cask.cdap;
+
+public class MiscTest {
+
+  public static void main(String[] args) {
+
+  }
+}